TestNinja/Mocking/EmployeeController.cs
using Microsoft.EntityFrameworkCore;
using TestNinja.Mocking.Refactored;

namespace TestNinja.Mocking
{
    public class EmployeeController
    {
        private readonly IEmployeeRepository _repo;

        public EmployeeController(IEmployeeRepository repo)
        {
            _repo = repo ?? new EmployeeRepository();
        }

        public ActionResult DeleteEmployee(int id)
        {
            _repo.Delete(id);
            return RedirectToAction("Employees");
        }

        private ActionResult RedirectToAction(string employees)
        {
            return new RedirectResult();
        }
    }

    public class ActionResult { }
 
    public class RedirectResult : ActionResult { }
    
    public class EmployeeContext : DbContext
    {
        public DbSet<Employee> Employees { get; set; }
    }

    public class Employee
    {
    }
}